Distance is now double

+++ src/main/java/com/martiply/model/interfaces/IStore.java
@@ -20,7 +20,7 @@ public interface IStore {
 
     public String getClose();
 
-    public float getDistance();
+    public double getDistance();
 
     public String getStory();
